+++ vice/src/monitor/mon_disassemble.h
@@ -32,8 +32,8 @@
 #include "montypes.h"
 #include "types.h"
 
-extern const char *mon_disassemble_to_string_ex(MEMSPACE, WORD addr, BYTE x,
-                                             BYTE p1, BYTE p2, BYTE p3,
+extern const char *mon_disassemble_to_string_ex(MEMSPACE, unsigned int addr, unsigned int x,
+                                             unsigned int p1, unsigned int p2, unsigned int p3,
                                              int hex_mode, unsigned *len);
 extern void mon_disassemble_lines(MON_ADDR start_addr, MON_ADDR end_addr);
